- Altered effects to not update when they're attached to objects which are in an unloaded sector.

+++ src/common/net/netbase.h
@@ -374,7 +374,7 @@ protected:
         struct timeval prevTimeout = timeout;
 
         /* select returns 0 if timeout, 1 if input available, -1 if error. */
-        if (SOCK_SELECT(MAX(mysocket, pipe_fd[0]) + 1, &set, NULL, NULL, &timeout) < 1)
+        if (SOCK_SELECT(MAX(mysocket, (uint)pipe_fd[0]) + 1, &set, NULL, NULL, &timeout) < 1)
         {
             timeout = prevTimeout;
             return 0;
